# apps/simple-agents/src/doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Simple Agents API",
    "description": "AI agent session management API with chat, streaming, and artifact support.\n\n## Streaming Endpoints\n\nTwo streaming modes are available:\n\n- **`/chat/stream`** (AI SDK): Uses `@ai-sdk/langchain` `toUIMessageStream()` to output standard AI SDK `UIMessageChunk` events. Recommended for frontend `useChat` integration.\n- **`/chat/raw-stream`** (Raw): Manually iterates `streamEvents()` outputting custom `StreamEvent` objects. Preserves full LangChain event granularity. Supports disconnect recovery via `/stream/{streamId}`.",
    "version": "2.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/sessions/{id}": {
      "get": {
        "operationId": "getSession",
        "summary": "Get a session by ID",
        "tags": ["Sessions"],
        "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
        "responses": {
          "200": {
            "description": "Session found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateSession",
        "summary": "Update a session",
        "tags": ["Sessions"],
        "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateSessionInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteSession",
        "summary": "Delete a session",
        "tags": ["Sessions"],
        "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
        "responses": {
          "200": {
            "description": "Session deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/messages": {
      "get": {
        "operationId": "getMessages",
        "summary": "List messages in a session",
        "tags": ["Chat"],
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1, "minimum": 1 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer", "default": 50, "minimum": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated message list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMessageResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sessions/{id}/stream": {
      "get": {
        "operationId": "getStreamStatus",
        "summary": "Get the current stream status for a session",
        "description": "Only applicable to raw-stream sessions. Checks the in-memory stream registry first, then falls back to the database.",
        "tags": ["Streaming"],
        "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
        "responses": {
          "200": {
            "description": "Stream status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StreamStatus" }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/stream/{streamId}": {
      "get": {
        "operationId": "resumeStream",
        "summary": "Resume or retrieve a raw stream by ID",
        "description": "Only applicable to raw-stream sessions. If the stream is still running, returns an SSE continuation. If completed, returns JSON with the result. Falls back to the database for page-reload recovery.",
        "tags": ["Streaming"],
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" },
          { "$ref": "#/components/parameters/StreamId" }
        ],
        "responses": {
          "200": {
            "description": "Stream result or SSE continuation stream",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StreamResult" }
              },
              "text/event-stream": {
                "schema": {
                  "description": "SSE stream resuming from cached content or live events."
                }
              }
            }
          },
          "404": {
            "description": "Stream not found or expired",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StreamNotFound" }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/stream/{streamId}/cancel": {
      "post": {
        "operationId": "cancelStream",
        "summary": "Cancel an active raw stream",
        "description": "Only applicable to raw-stream sessions. Aborts the background agent task and stops the SSE connection.",
        "tags": ["Streaming"],
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" },
          { "$ref": "#/components/parameters/StreamId" }
        ],
        "responses": {
          "200": {
            "description": "Stream cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "streamId": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Stream not found, already completed, or expired",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "SessionId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Session ID"
      },
      "StreamId": {
        "name": "streamId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Stream task ID"
      },
...
    },
    "schemas": {
...
      "Session": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "metadata": { "type": ["string", "null"] },
          "isArchived": { "type": "boolean" }
        },
        "required": ["id", "title", "createdAt", "updatedAt", "isArchived"]
      },
...
      "UpdateSessionInput": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "metadata": { "type": "object" },
          "isArchived": { "type": "boolean" }
        }
      },
      "Message": {
        "type": "object",
        "description": "A stored message. The `parts` field contains the full AI SDK UIMessage parts as a JSON array (for assistant messages created via the AI SDK stream endpoint). The `content` field is a plain-text fallback.",
        "properties": {
          "id": { "type": "integer" },
          "sessionId": { "type": "string" },
          "role": {
            "type": "string",
            "enum": ["user", "assistant", "system", "tool"]
          },
          "content": {
            "type": ["string", "null"],
            "description": "Plain text content (fallback)"
          },
          "parts": {
            "type": ["string", "null"],
            "description": "JSON-encoded UIMessage parts array. Present for messages created via the AI SDK stream endpoint. Contains text, tool-invocation, tool-result parts etc."
          },
          "toolCalls": {
            "type": ["string", "null"],
            "description": "JSON-encoded tool calls (legacy)"
          },
          "toolResults": {
            "type": ["string", "null"],
            "description": "JSON-encoded tool results (legacy)"
          },
          "tokenCount": { "type": ["integer", "null"] },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "sessionId", "role", "createdAt"]
      },
...
      "StreamStatus": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "active": { "type": "boolean", "const": true },
              "streamId": { "type": "string" },
              "status": { "type": "string" }
            }
          },
          {
            "type": "object",
            "properties": {
              "active": { "type": "boolean", "const": false },
              "lastStreamId": { "type": "string" },
              "lastStatus": { "type": "string" },
              "lastMessageId": { "type": "integer" }
            }
          }
        ],
        "discriminator": { "propertyName": "active" }
      },
      "StreamResult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["completed", "failed", "cancelled"]
          },
          "content": { "type": "string" },
          "error": { "type": ["string", "null"] },
          "messageId": { "type": ["integer", "null"] },
          "streamId": { "type": "string" },
          "source": { "type": ["string", "null"] }
        }
      },
      "StreamNotFound": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "latestStreamId": { "type": "string" },
          "latestStatus": { "type": "string" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
...
      "PaginatedMessageResult": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Message" }
          },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      }
    }
  },
...
}